pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.48.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
]
